+++ keep_screen_on_platform_interface/test/method_channel_keep_screen_on_test.dart
@@ -1,3 +1,4 @@
+
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:keep_screen_on_platform_interface/keep_screen_on_platform_interface.dart';
@@ -15,20 +16,26 @@ void main() {
     methodChannel = const MethodChannel('dev.craftsoft/keep_screen_on');
     keepScreenOn = MethodChannelKeepScreenOn.private(methodChannel);
 
-    methodChannel.setMockMethodCallHandler((call) async {
-      log.add(call);
+    TestDefaultBinaryMessengerBinding
+        .instance
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(methodChannel, (message) {
+          log.add(message);
 
-      switch (call.method) {
-        case 'isOn':
-          return null;
+          switch (message.method) {
+            case 'isOn':
+              return null;
 
-        case 'isOff':
-          return null;
+            case 'isOff':
+              return null;
 
-        default:
-          return null;
-      }
-    });
+            case 'isAllowLockWhileScreenOn':
+              return null;
+
+            default:
+              return null;
+          }
+        });
 
     log.clear();
   });
@@ -37,17 +44,27 @@ void main() {
     
     test('turnOn', () async {
       await keepScreenOn.turnOn();
-      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'turnOn': true})]);
+      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'on': true, 'withAllowLockWhileScreenOn': false})]);
     });
 
-    test('turnOn false', () async {
-      await keepScreenOn.turnOn(false);
-      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'turnOn': false})]);
+    test('turnOn(on: false)', () async {
+      await keepScreenOn.turnOn(on: false);
+      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'on': false, 'withAllowLockWhileScreenOn': false})]);
+    });
+
+    test('turnOn(on: true, withAllowLockWhileScreenOn: true)', () async {
+      await keepScreenOn.turnOn(on: true, withAllowLockWhileScreenOn: true);
+      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'on': true, 'withAllowLockWhileScreenOn': true})]);
     });
 
     test('turnOff', () async {
       await keepScreenOn.turnOff();
-      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'turnOn': false})]);
+      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'on': false, 'withAllowLockWhileScreenOn': false})]);
+    });
+
+    test('turnOff(withAllowLockWhileScreenOn: true)', () async {
+      await keepScreenOn.turnOff(withAllowLockWhileScreenOn: true);
+      expect(log, <Matcher>[isMethodCall('turnOn', arguments: <String, Object>{'on': false, 'withAllowLockWhileScreenOn': true})]);
     });
 
     test('isOn', () async {
@@ -60,6 +77,26 @@ void main() {
       final isOff = await keepScreenOn.isOff;
       expect(isOff, null);
       expect(log, <Matcher>[isMethodCall('isOn', arguments: null)]);
+    });
+
+    test('isAllowLockWhileScreenOn', () async {
+      final isAllowLockWhileScreenOn = await keepScreenOn.isAllowLockWhileScreenOn;
+      expect(isAllowLockWhileScreenOn, null);
+    });
+
+    test('addAllowLockWhileScreenOn', () async {
+      final result = await keepScreenOn.addAllowLockWhileScreenOn();
+      expect(result, false);
+    });
+
+    test('addAllowLockWhileScreenOn(on: false)', () async {
+      final result = await keepScreenOn.addAllowLockWhileScreenOn(on: false);
+      expect(result, false);
+    });
+
+    test('clearAllowLockWhileScreenOn', () async {
+      final result = await keepScreenOn.clearAllowLockWhileScreenOn();
+      expect(result, false);
     });
   });
 }